% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{pressR-package}
\alias{pressR}
\alias{pressR-package}
\title{pressR: Scrape Press Releases from U.S. House Members' Websites}
\description{
Collects press releases from U.S. House of Representatives members'
\code{.house.gov} sites over a date range. The package detects the
content-management system behind each site and routes to a dedicated
extractor. See \code{\link[=scrape_pressers]{scrape_pressers()}} for the main entry point.
}
\author{
\strong{Maintainer}: Zayne Sember \email{zaynesember@gmail.com}

}
\keyword{internal}
